Added SmartMove AI
it's fairly working ai though need's quite alot improvement's
for it to work properly but right now it does it's job good
enough to consider it working.

it's designed for snakes to follow food and eat it..
if there's alot of food (more than 10) snake's will use it's
boost ability.